Port X navs w/ sidebar
